libcamera: base: signal: Disallow rvalue references
The main use of rvalue references is "moving", usually taking
ownership of the resources of an object and setting it to some
kind of "empty" state. However, it is not clear how that could
(or should) work with a signal that has multiple recipients,
so disallow the use of rvalue references as signal arguments.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>