libcamera: base: bound_method: Forward when invoking
This implements the same kind of change for pack-based invoking as 9c86a405b72a
("libcamera: base: bound_method: Forward arguments when possible") did for the
other cases.

At the moment a single argument pack is used just once, hence it is safe
to forward the arguments from the pack, allowing moving to take place
instead of copying. So do that.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>