meson: Bump minimum version to v1.0.1
The minimum version of meson currently required by libcamera, v0.63,
does not properly detect Qt6. As qcam switched from Qt5 to Qt6 a while
ago, building libcamera with meson v0.63 is currently broken.

The issue was fixed in meson v0.64. That was the last meson version
before v1.0.0. For mainly psychological reasons, and to avoid bumping
the meson version too often, go straight to v1.0.1. This will support
the following versions of popular distributions and build environments:

- Debian Bookworm and newer (v1.0.1)
- Ubuntu 24.04LTS and newer (v1.3.2)
- Buildroot 2023.08 (v1.1.1)
- Yocto Scarthgap (v1.3.1)

The previous version of Debian, Ubuntu and Yocto shipped versions of
meson older than v0.63 (v0.56.2 in Debian Bullseye, v0.61.2 in Ubuntu
22.04LTS, and v0.61.3 in Yocto Kirkstone), so the meson version bump
doesn't require a distribution upgrade. For Buildroot, moving to v1.0.1
means users on versions from 2022.08 (when not using qcam) or 2023.02 to
2023.05 will need to upgrade.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>